# backend/src/main/java/com/studentinfo/dto/AuthResponse.java
package com.studentinfo.dto;

import com.studentinfo.model.User;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class AuthResponse {
    private String token;
    private User user;
}