# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Student Information System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /departments:
    get:
      operationId: listDepartments
      tags: [departments]
      summary: List all departments
      responses:
        "200":
          description: List of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [departments]
      summary: Create a department (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentInput"
      responses:
        "201":
          description: Department created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
  /departments/{departmentId}:
    get:
      operationId: getDepartment
      tags: [departments]
      summary: Get department by ID
      parameters:
        - name: departmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Department found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
    patch:
      operationId: updateDepartment
      tags: [departments]
      summary: Update a department
      parameters:
        - name: departmentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentUpdate"
      responses:
        "200":
          description: Department updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
    delete:
      operationId: deleteDepartment
      tags: [departments]
      summary: Delete a department
      parameters:
        - name: departmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /students/{studentId}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get student by ID
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Delete student record
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard statistics (role-filtered)
      parameters:
        - name: departmentId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: ["string", "null"]
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, email, role, fullName]
      properties:
        id:
          type: integer
        email:
          type: string
        role:
          type: string
          enum: [SUPER_ADMIN, CENTRAL_PLACEMENT, DEPT_PLACEMENT, SRC, STUDENT]
        fullName:
          type: string
        departmentId:
          type: ["integer", "null"]
        departmentName:
          type: ["string", "null"]
        studentId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
...
    Department:
      type: object
      required: [id, name, code]
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        headName:
          type: ["string", "null"]
        studentCount:
          type: integer
        placedCount:
          type: integer
    DepartmentInput:
      type: object
      required: [name, code]
      properties:
        name:
          type: string
        code:
          type: string
        headName:
          type: ["string", "null"]
    DepartmentUpdate:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        headName:
          type: ["string", "null"]
    Student:
      type: object
      required: [id, rollNumber, userId, departmentId, departmentName, fullName]
      properties:
        id:
          type: integer
        rollNumber:
          type: string
        userId:
          type: integer
        fullName:
          type: string
        email:
          type: string
        departmentId:
          type: integer
        departmentName:
          type: string
        year:
          type: integer
        section:
          type: ["string", "null"]
        placementStatus:
          type: string
          enum: [NOT_PLACED, PLACED, OPTED_OUT]
        profileCompletionPercent:
          type: integer
    StudentProfile:
      type: object
      required: [student]
      properties:
        student:
          $ref: "#/components/schemas/Student"
        personal:
          $ref: "#/components/schemas/PersonalDetails"
        academic:
          $ref: "#/components/schemas/AcademicDetails"
        experience:
          type: array
          items:
            $ref: "#/components/schemas/ExperienceEntry"
        skills:
          $ref: "#/components/schemas/SkillsDetails"
        achievements:
          type: array
          items:
            $ref: "#/components/schemas/Achievement"
...
    PersonalDetails:
      type: object
      required: [studentId]
      properties:
        studentId:
          type: integer
        dateOfBirth:
          type: ["string", "null"]
          format: date
        gender:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        alternatePhone:
          type: ["string", "null"]
        permanentAddress:
          type: ["string", "null"]
        currentAddress:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        pincode:
          type: ["string", "null"]
        nationality:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        aadharNumber:
          type: ["string", "null"]
        panNumber:
          type: ["string", "null"]
        linkedinUrl:
          type: ["string", "null"]
        githubUrl:
          type: ["string", "null"]
        portfolioUrl:
          type: ["string", "null"]
...
    AcademicDetails:
      type: object
      required: [studentId]
      properties:
        studentId:
          type: integer
        tenthBoard:
          type: ["string", "null"]
        tenthSchool:
          type: ["string", "null"]
        tenthPercentage:
          type: ["number", "null"]
        tenthYear:
          type: ["integer", "null"]
        twelfthBoard:
          type: ["string", "null"]
        twelfthSchool:
          type: ["string", "null"]
        twelfthPercentage:
          type: ["number", "null"]
        twelfthYear:
          type: ["integer", "null"]
        diplomaInstitute:
          type: ["string", "null"]
        diplomaPercentage:
          type: ["number", "null"]
        diplomaYear:
          type: ["integer", "null"]
        ugInstitute:
          type: ["string", "null"]
        ugDegree:
          type: ["string", "null"]
        ugBranch:
          type: ["string", "null"]
        ugCgpa:
          type: ["number", "null"]
        ugPassingYear:
          type: ["integer", "null"]
        currentSemesterCgpa:
          type: ["number", "null"]
        activeBacklogs:
          type: ["integer", "null"]
        totalBacklogs:
          type: ["integer", "null"]
        gapYears:
          type: ["integer", "null"]
...
    ExperienceEntry:
      type: object
      required: [id, studentId, title, company, type]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        type:
          type: string
          enum: [INTERNSHIP, FULL_TIME, PART_TIME, FREELANCE, PROJECT]
        title:
          type: string
        company:
          type: string
        location:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
          format: date
        endDate:
          type: ["string", "null"]
          format: date
        isCurrent:
          type: boolean
        description:
          type: ["string", "null"]
        stipend:
          type: ["number", "null"]
        offerLetter:
          type: boolean
        completionCertificate:
          type: boolean
...
    SkillsDetails:
      type: object
      required: [studentId]
      properties:
        studentId:
          type: integer
        technicalSkills:
          type: array
          items:
            type: string
        softSkills:
          type: array
          items:
            type: string
        languages:
          type: array
          items:
            type: string
        tools:
          type: array
          items:
            type: string
        certifications:
          type: array
          items:
            $ref: "#/components/schemas/Certification"
...
    Certification:
      type: object
      required: [id, name, issuer]
      properties:
        id:
          type: integer
        name:
          type: string
        issuer:
          type: string
        issueDate:
          type: ["string", "null"]
          format: date
        expiryDate:
          type: ["string", "null"]
          format: date
        credentialId:
          type: ["string", "null"]
        credentialUrl:
          type: ["string", "null"]
...
    Achievement:
      type: object
      required: [id, studentId, title]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        date:
          type: ["string", "null"]
          format: date
        category:
          type: ["string", "null"]
        proof:
          type: ["string", "null"]
...
    DashboardStats:
      type: object
      required: [totalStudents, placedStudents, activeDrives, departmentsCount]
      properties:
        totalStudents:
          type: integer
        placedStudents:
          type: integer
        activeDrives:
          type: integer
        departmentsCount:
          type: integer
        placementPercentage:
          type: number
        avgPackage:
          type: ["string", "null"]
        profileCompletionAvg:
          type: number
        totalUsers:
          type: integer
...